+++ src/tracker/stack/definition.ts
@@ -1,7 +1,8 @@
-import type { InvocationKey } from './invocation-tracker.ts';
+import type { InvocationKey } from '../definition.ts';
 
 /**
- * A stack abstraction used by the invocation tracker to track the nesting of invocation keys.
+ * A stack abstraction used by the invocation tracker (`createInvocationTracker`) to track the nesting of invocation
+ * keys.
  *
  * `InvocationStack` enables the tracker to determine parent-child relationships between invocations by tracking the
  * currently active `InvocationKey` at any point in the program. This is particularly useful when correlating nested or
@@ -10,13 +11,6 @@ import type { InvocationKey } from './invocation-tracker.ts';
  * The tracker uses the top value from the stack (via `peek()`) to determine the parent for the next invocation. The
  * stack is updated automatically by the tracker: `push()` is called at the start of a tracked function, and `pop()` is
  * called after the function completes or throws. The return value of `pop()` is the key that was just removed.
- *
- * There are two built-in implementations:
- *
- * - `createInvocationStack()` — a synchronous, in-memory implementation (default). Suitable for browser or
- *   single-threaded environments.
- * - `createThreadSafeInvocationStack()` — a Node.js implementation that uses `AsyncLocalStorage` to maintain stack state
- *   across async boundaries.
  *
  * Clients may provide their own implementation if they need integration with custom async contexts, request-local
  * storage, or thread-local semantics in other platforms. Any implementation must conform to this interface.
@@ -47,32 +41,10 @@ export type InvocationStack = {
 };
 
 /**
- * A stack that tracks the nesting of invocation keys without support for multithread scenarios.
- *
- * This stack is used to track parent-child relationships between invocations by pushing the `InvocationKey` of each
- * active function call and removing it after completion.
- *
- * This implementation is not thread-safe and is suitable for environments where each thread or event loop has its own
- * isolated context (e.g., browser, single-threaded Node).
- *
- * This is the default invocation stack used when creating an invocation tracker (`createInvocationTracker`).
- *
- * @returns A synchronous, in-memory invocation stack.
+ * A stack storage to support thread-safe scenarios, modeled after NodeJS's `AsyncLocalStorage`.
  */
-export const createInvocationStack = (): InvocationStack => {
-  const stack: InvocationKey[] = [];
-
-  return {
-    close: () => {
-      stack.length = 0;
-    },
-
-    push: (key: InvocationKey) => {
-      stack.push(key);
-    },
-
-    peek: () => stack.at(-1),
-
-    pop: () => stack.pop(),
-  };
-};
+export interface AsyncStackStorage {
+  getStore(): InvocationKey[] | undefined;
+  enterWith(store: InvocationKey[]): void;
+  disable(): void;
+}
